Move mitm code to own package
